refactor: moved services from 'lib/util' to 'lib/services'
