+++ apps/cli/src/lib/services/BrowserService.ts
