+++ apps/cli/src/lib/services/GlobalConfigService.ts
