+++ apps/cli/src/modules/auth/auth.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ApiService, BrowserService, GlobalConfigService } from '../../lib/util';
+import { ApiService, BrowserService, GlobalConfigService } from '../../lib/services';
 import { Login } from './login/login.command';
 import { LoginQuestions } from './login/login.questions';
